+++ src/pandora_graphic.h
@@ -38,7 +38,7 @@ void update_graph_data(type_group *group);
 
 void graph_get_line_color(int num, float *r, float *g, float *b);
 //void group_update_display(TxGraphic *cr, int a, int b, int c, int d, type_group *g, int y_offset,  int z, int zMax);
-void architecture_set_view_point(GtkWidget *scrollbars, float x, float y);
+void architecture_set_view_point(GtkWidget *scrollbars, gdouble x, gdouble y);
 void architecture_get_group_position(type_group *group, float *x, float *y);
 void architecture_display_update(TxWidget *group_display, void *event);
 void architecture_display_update_group(GtkWidget *architecture_display, type_group *group);
